refactor(mux): rehome the cross-backend live-session snapshot
Move LiveSessions and its precedence test intact beside backend selection.
Room inventory, reload, remote control, and web borrow the mux-owned
snapshot; reload and remote control no longer depend on room.
