+++ crates/rimz/src/mux/AGENTS.md
@@ -13,6 +13,7 @@ Local contract for `crates/rimz/src/mux/` — the Zellij/tmux seam. Extends [cra
 
 ## Ownership and direction
 
+- `LiveSessions` in `selection.rs` owns the one batch snapshot of both backends' live sessions: errors read as empty and session resolution prefers Zellij. Room inventory, reload, remote control, and web borrow it rather than re-deriving it.
 - Presence-plugin lifecycle verbs live in `zellij/presence.rs` and `zellij/session.rs`: load-if-missing through `MuxBackend::ensure_presence_plugin`, explicit convergence, reload upgrade with its `PresenceUpgrade` outcome, topology dump, and the fresh-topology wait. Reload consumes outcomes and owns its counters and logs; `PresencePluginOptions::from_config` in `mod.rs` is the one config-to-options projection.
 - `mux` never imports `sidebar`. The wire every renderer listens on (the heartbeat record, the event vocabulary, the datagram send) is the leaf [`wakeup/`](../wakeup/mod.rs) module below `store`; every sender reaches down to it (the store's write tail, `mux`, `remote_control`, the sidebar graph and its renderer, the CLI), and none reaches another module through it.
 - Every runtime file the multiplexer writes or dispatches on is owned here: the durable focus intent behind each RimZ-initiated jump in [`focus_anchor.rs`](./focus_anchor.rs), the room-wide width target in [`width_target.rs`](./width_target.rs), the Zellij pane-topology and presence-desired caches in [`zellij/pane_topology.rs`](./zellij/pane_topology.rs), and the per-pane advisory write lock in [`pane_writer.rs`](./pane_writer.rs), whose user-scoped runtime path comes from `RuntimePaths::pane_write_lock`. Their path, schema, and freshness rules stay here even where the writer sits elsewhere: the sidebar's Zellij wake ingestion publishes the topology cache through these helpers rather than spelling the file itself.
